uploaded work done so far
this commit includes all the work that have been done in the project.
All the work here is a work in progress and a matter of pure experiment.
These were uploaded here for work collaboration with other team members
of the project. These codes are not for use by anyone anywhere except
for the project development purposes of our team.
we will be up with the working version soon!! :)
Stay tuned!!